+++ test/private/deconstructJsdocNamepath.test.js
@@ -15,13 +15,21 @@ module.exports = (tests) => {
   tests.add(
     '`deconstructJsdocNamepath` with nested static, instance and inner members.',
     () => {
-      deepStrictEqual(deconstructJsdocNamepath('a.b#c~d'), {
-        memberof: 'a.b#c',
+      deepStrictEqual(deconstructJsdocNamepath('a.B#c~d'), {
+        memberof: 'a.B#c',
         membership: '~',
         name: 'd',
       });
     }
   );
+
+  tests.add('`deconstructJsdocNamepath` with a namepath prefix.', () => {
+    deepStrictEqual(deconstructJsdocNamepath('A#event:a'), {
+      memberof: 'A',
+      membership: '#',
+      name: 'event:a',
+    });
+  });
 
   tests.add('`deconstructJsdocNamepath` with invalid namepaths.', () => {
     for (const namepath of ['', 'a..b', 'a..b.c', 'a.'])